+++ services/workflows-service/src/workflow/types/params.ts
@@ -4,9 +4,3 @@ export interface FindLastActiveFlowParams {
   workflowDefinitionId: string;
   businessId: string;
 }
-
-export interface GetLastActiveFlowParams {
-  email: string;
-  workflowRuntimeDefinitionId: string;
-  projectIds: TProjectIds;
-}