+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/headless-example
 
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.45
+  - @ballerine/workflow-browser-sdk@0.5.45
+
+## 0.1.44
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.44
+  - @ballerine/workflow-browser-sdk@0.5.44
+
 ## 0.1.43
 
 ### Patch Changes