+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ballerine/workflow-core
 
+## 0.5.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.45
+
+## 0.5.44
+
+### Patch Changes
+
+- Fixes for concurrency fixes
+- Updated dependencies
+  - @ballerine/common@0.7.44
+
 ## 0.5.43
 
 ### Patch Changes